// src/types/template.ts
export interface OceaniaAboutProps {
    badgeText: string;
    title: string;
    description: string;
    mission?: string;
    vision?: string;
    values?: string[];
    purposeTitle: string;
    purposeDescription: string;
}